Document cooldown orchestrator helpers
